+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 12.0.1
 - Fix no REMOVE column (add button) if no periods in the list in Periods.php
 - Fix SQL error when no Periods, make Period select required in Courses.php
 - Security fix #352 selected theme does not exist in Preferences.php, thanks to @rudranshsinghrajpurohit
+- Security fix selected page size does not exist in Preferences.php
 
 Changes in 12.0
 ---------------