+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 12.0.1
 - Security fix selected page size does not exist in Preferences.php
 - Remove "-main" suffix from manually uploaded add-ons in Modules.inc.php & Plugins.inc.php
 - Fix PHP error when Recreate Calendar & From or To date is N/A in Calendar.php
+- Add checkout latest release tag to installation instrcutions in INSTALL.md, INSTALL_es.md & INSTALL_fr.md
 
 Changes in 12.0
 ---------------