+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 12.0.1
 - Remove "-main" suffix from manually uploaded add-ons in Modules.inc.php & Plugins.inc.php
 - Fix PHP error when Recreate Calendar & From or To date is N/A in Calendar.php
 - Add checkout latest release tag to installation instrcutions in INSTALL.md, INSTALL_es.md & INSTALL_fr.md
+- Fix PHP warning undefined array key "PROFILE_ID" in GetTeacher.fnc.php
 
 Changes in 12.0
 ---------------